Sets password public and secret fingerprints.
Instead of setting the password just for the secret key's fingerprint,
set the password for the public key as well.  It appears that GPG has an
inconsistency where `gpg -d` looks up the key using the secret key's
finterprint, but, for example, `reprepro` uses the public key's
fingerprint for signing the APT repository files.